Split animation mixer to Mixer file
Cleanup

Remove linq

Fix

Fixes
