+++ Renderer/Renderer/AnimationController.Mixer.cs
@@ -1,0 +1,388 @@
+using System.Diagnostics;
+using ValveResourceFormat.ResourceTypes.ModelAnimation;
+
+namespace ValveResourceFormat.Renderer
+{
+    /// <summary>
+    /// Manages skeletal animation playback and computes animated bone poses.
+    /// </summary>
+    public partial class AnimationController
+    {
+        /// <summary>Represents an animation clip with its playback state.</summary>
+        public record class Clip(Animation Animation)
+        {
+            /// <summary>Gets or sets the current playback time in seconds.</summary>
+            public float Time { get; set; }
+
+            /// <summary>Gets or sets whether playback is paused.</summary>
+            public bool IsPaused { get; set; }
+
+            /// <summary>Gets or sets whether the clip should loop when reaching the end.</summary>
+            public bool Looping { get; set; } = true;
+
+            /// <summary>Gets or sets the blend weight (0.0 to 1.0) for this clip.</summary>
+            public float Weight { get; set; } = 1f;
+
+            /// <summary>Gets or sets the blend transition time in seconds. Negative values indicate manual blending.</summary>
+            public float BlendTime { get; set; }
+
+            /// <summary>Gets whether this clip uses time-based transition blending.</summary>
+            public bool IsTimeBasedTransition => BlendTime > 0f;
+
+            /// <summary>Gets whether this clip uses manual weight blending.</summary>
+            public bool IsManualBlend => BlendTime == -1;
+
+            /// <summary>Gets or sets the current frame index.</summary>
+            public int Frame
+            {
+                get
+                {
+                    if (Animation.FrameCount > 1)
+                    {
+                        return (int)MathF.Round(Time * Animation.Fps) % Animation.FrameCount;
+                    }
+                    return 0;
+                }
+                set
+                {
+                    Time = Animation.Fps != 0 ? value / Animation.Fps : 0f;
+                }
+            }
+        }
+
+        /// <summary>
+        /// Gets whether the active animation clip has finished playing (is not looping and has reached the end).
+        /// </summary>
+        public bool ActiveClipFinished => CurrentSubController.HasValue
+            ? CurrentSubController.Value.Handler.ActiveClipFinished
+            : activeClip != null && !activeClip.Looping && activeClip.IsPaused;
+
+        private Clip? activeClip;
+        private Clip? previousClip;
+        private readonly Dictionary<string, Clip> clips = [];
+        private readonly Frame BlendedFrame;
+        private float currentBlendTime;
+
+        /// <summary>
+        /// Updates time and weights for all active clips during playback.
+        /// </summary>
+        /// <param name="timeStep">Elapsed time in seconds since the last update.</param>
+        private void UpdateClips(float timeStep)
+        {
+            if (activeClip == null)
+            {
+                return;
+            }
+
+            var allPaused = true;
+            foreach (var clip in clips.Values)
+            {
+                if (!clip.IsPaused)
+                {
+                    allPaused = false;
+                    break;
+                }
+            }
+            IsPaused = allPaused;
+
+            // Update time for all clips
+            foreach (var clip in clips.Values)
+            {
+                if (!clip.IsPaused && clip.Animation.FrameCount > 1)
+                {
+                    clip.Time += timeStep;
+
+                    if (!clip.Looping)
+                    {
+                        var lastFrame = clip.Animation!.FrameCount - 1;
+                        var maxTime = lastFrame / clip.Animation.Fps;
+
+                        if (clip.Time > maxTime)
+                        {
+                            clip.IsPaused = true;
+                            clip.Frame = lastFrame;
+                        }
+                    }
+                }
+            }
+
+            // Sync global Time and Frame with active clip
+            Time = activeClip.Time;
+            Frame = activeClip.Frame;
+
+            if (activeClip.IsTimeBasedTransition && previousClip != null)
+            {
+                // Distribute blend weights between previous clip and active clip only.
+                currentBlendTime -= timeStep;
+
+                if (currentBlendTime <= 0f)
+                {
+                    previousClip.Weight = 0f;
+                    activeClip.Weight = 1f;
+                    previousClip = null;
+                }
+                else
+                {
+                    var t = activeClip.BlendTime > 0f
+                        ? 1f - Math.Clamp(currentBlendTime / activeClip.BlendTime, 0f, 1f)
+                        : 1f;
+
+                    var blendProgress = t * t * (3f - 2f * t);
+
+                    activeClip.Weight = blendProgress;
+                    previousClip.Weight = 1f - blendProgress;
+
+                    foreach (var clip in clips.Values)
+                    {
+                        if (clip != activeClip && clip != previousClip)
+                        {
+                            clip.Weight = 0f;
+                        }
+                    }
+                }
+
+                var sum = 0f;
+                foreach (var clip in clips.Values)
+                {
+                    sum += clip.Weight;
+                }
+                Debug.Assert(sum > 0f, "Total blend weight should be greater than zero.");
+                Debug.Assert(Math.Abs(sum - 1f) < 0.01f, $"Total blend weight should be approximately 1. Found: {sum}");
+            }
+        }
+
+        /// <summary>
+        /// Returns the animation frame for the current time, blending multiple clips if needed.
+        /// </summary>
+        /// <returns>The current animation frame, or <see langword="null"/> if no animation is active.</returns>
+        private Frame? GetBlendedFrame()
+        {
+            if (activeClip == null)
+            {
+                return null;
+            }
+
+            // Check if blending is needed
+            var needsBlending = false;
+            foreach (var clip in clips.Values)
+            {
+                if (clip != activeClip && clip.Weight > 0f)
+                {
+                    needsBlending = true;
+                    break;
+                }
+            }
+
+            if (!needsBlending)
+            {
+                return SampleFrame(activeClip);
+            }
+
+            BlendedFrame.FrameIndex = -1;
+            BlendedFrame.Bones.AsSpan().Clear();
+            BlendedFrame.Datas.AsSpan().Clear();
+
+            var totalWeight = 0f;
+            foreach (var clip in clips.Values)
+            {
+                if (clip.Weight <= 0f)
+                {
+                    continue;
+                }
+
+                var frame = SampleFrame(clip);
+                var blendFactor = clip.Weight / (totalWeight + clip.Weight);
+
+                for (var i = 0; i < frame.Bones.Length; i++)
+                {
+                    BlendedFrame.Bones[i] = BlendedFrame.Bones[i].Blend(frame.Bones[i], blendFactor);
+                }
+
+                for (var i = 0; i < frame.Datas.Length; i++)
+                {
+                    BlendedFrame.Datas[i] = float.Lerp(BlendedFrame.Datas[i], frame.Datas[i], blendFactor);
+                }
+
+                totalWeight += clip.Weight;
+            }
+
+            return BlendedFrame;
+        }
+
+        private Frame SampleFrame(Clip clip)
+        {
+            var ignoreCache = clip.Animation != ActiveAnimation;
+
+            try
+            {
+                if (ignoreCache)
+                {
+                    FrameCache.PurgeCache();
+                }
+
+                return clip.IsPaused
+                    ? FrameCache.GetFrame(clip.Animation, clip.Frame)
+                    : FrameCache.GetInterpolatedFrame(clip.Animation, clip.Time);
+            }
+            finally
+            {
+                if (ignoreCache)
+                {
+                    FrameCache.PurgeCache();
+                }
+            }
+        }
+
+        /// <summary>
+        /// Transitions to a new animation clip with the specified blend time, managing clip weights appropriately.
+        /// </summary>
+        /// <param name="animation">The animation to transition to.</param>
+        /// <param name="blendTime">The blend time in seconds. 0 for instant transition, -1 for manual blending.</param>
+        private void TransitionToClip(Animation animation, float blendTime)
+        {
+            var animName = animation.Name;
+
+            // Check if clip already exists
+            if (!clips.TryGetValue(animName, out var newClip))
+            {
+                newClip = new Clip(animation) { Looping = Looping, BlendTime = blendTime };
+                clips[animName] = newClip;
+            }
+            else
+            {
+                // Update existing clip properties
+                newClip.Looping = Looping;
+                newClip.BlendTime = blendTime;
+
+                newClip.IsPaused = false;
+                newClip.Frame = 0;
+            }
+
+            // Handle blending
+            if (activeClip == newClip)
+            {
+                // Re-setting the same animation should not create a self-blend transition.
+                previousClip = null;
+
+                foreach (var clip in clips.Values)
+                {
+                    clip.Weight = 0f;
+                }
+
+                newClip.Weight = 1f;
+
+                if (blendTime == 0f)
+                {
+                    FrameCache.Clear();
+                }
+            }
+            else if (blendTime > 0f && activeClip != null)
+            {
+                // Time-based transition: only blend from previous clip -> active clip.
+                previousClip = activeClip;
+                previousClip.Weight = 1f;
+
+                // Set all other clips to zero immediately.
+                foreach (var clip in clips.Values)
+                {
+                    if (clip != previousClip && clip != newClip)
+                    {
+                        clip.Weight = 0f;
+                    }
+                }
+
+                newClip.Weight = 0f;
+                currentBlendTime = blendTime;
+            }
+            else if (blendTime == -1f && activeClip != null)
+            {
+                // Manual blend: keep previous clip, user may set weights manually.
+                previousClip = activeClip;
+                previousClip.Weight = 1f;
+
+                foreach (var clip in clips.Values)
+                {
+                    if (clip != previousClip && clip != newClip)
+                    {
+                        clip.Weight = 0f;
+                    }
+                }
+
+                newClip.Weight = 0f;
+            }
+            else
+            {
+                // No blending - disable previous clip and all other clips.
+                previousClip = null;
+
+                foreach (var clip in clips.Values)
+                {
+                    clip.Weight = 0f;
+                }
+
+                newClip.Weight = 1f;
+
+                if (blendTime == 0f)
+                {
+                    FrameCache.Clear();
+                }
+            }
+
+            activeClip = newClip;
+        }
+
+        /// <summary>
+        /// Sets the blend weight for a clip with the specified animation name.
+        /// </summary>
+        /// <param name="name">The name of the animation.</param>
+        /// <param name="weight">The weight value (0.0 to 1.0).</param>
+        /// <param name="restartIfNew">Whether to restart the animation if its just now fading in.</param>
+        public void SetAnimationWeight(string name, float weight, bool restartIfNew = false)
+        {
+            if (clips.TryGetValue(name, out var clip))
+            {
+                var wasZero = clip.Weight == 0f;
+                clip.Weight = weight;
+
+                if (restartIfNew && wasZero && weight > 0f)
+                {
+                    clip.Time = 0f;
+                    clip.IsPaused = false;
+                }
+            }
+
+            if (CurrentSubController is { } subController)
+            {
+                subController.Handler.SetAnimationWeight(name, weight, restartIfNew);
+            }
+        }
+
+        /// <summary>
+        /// Sets properties for a clip with the specified animation name.
+        /// </summary>
+        /// <param name="name">The name of the animation.</param>
+        /// <param name="time">Optional playback time to set.</param>
+        /// <param name="looping">Optional looping flag to set.</param>
+        public void SetAnimationProperties(string name, float? time = null, bool? looping = null)
+        {
+            if (clips.TryGetValue(name, out var clip))
+            {
+                if (time.HasValue)
+                {
+                    clip.Time = time.Value;
+                    clip.IsPaused = false;
+                }
+
+                if (looping.HasValue)
+                {
+                    clip.Looping = looping.Value;
+                }
+            }
+
+            if (CurrentSubController is { } subController)
+            {
+                subController.Handler.SetAnimationProperties(name, time, looping);
+            }
+        }
+    }
+}
