+++ Renderer/Renderer/AnimationController.cs
@@ -1,6 +1,3 @@
-using System.Diagnostics;
-using System.Linq;
-using System.Runtime.InteropServices;
 using ValveResourceFormat.ResourceTypes.ModelAnimation;
 using ValveResourceFormat.ResourceTypes.ModelFlex;
 
@@ -26,50 +23,10 @@ namespace ValveResourceFormat.Renderer
         public Matrix4x4 Transform { get; set; } = Matrix4x4.Identity;
 
         /// <summary>Gets or sets whether animations should loop when reaching the end.</summary>
-        public bool Looping { get; set; }
+        public bool Looping { get; set; } = true;
 
         /// <summary>Gets the currently active animation, or <see langword="null"/> if none is set.</summary>
         public Animation? ActiveAnimation { get; private set; }
-
-        /// <summary>Represents an animation clip with its playback state.</summary>
-        public record class Clip(Animation Animation)
-        {
-            public float Time { get; set; }
-            public bool IsPaused { get; set; }
-            public bool Looping { get; set; } = true;
-            public float Weight { get; set; } = 1f;
-            public float BlendTime { get; set; }
-
-            public bool IsTimeBasedTransition => BlendTime > 0f;
-            public bool IsManualBlend => BlendTime == -1;
-
-            public int Frame
-            {
-                get
-                {
-                    if (Animation.FrameCount > 1)
-                    {
-                        return (int)MathF.Round(Time * Animation.Fps) % Animation.FrameCount;
-                    }
-                    return 0;
-                }
-                set
-                {
-                    Time = Animation.Fps != 0 ? value / Animation.Fps : 0f;
-                }
-            }
-        }
-
-        private Clip? activeClip;
-        private Clip? previousClip;
-        private readonly Dictionary<string, Clip> clips = [];
-        private readonly Frame BlendedFrame;
-        private float currentBlendTime;
-
-        /// <summary>
-        /// Gets or sets the tilt-twist constraints that are applied when animations update.
-        /// </summary>
-        public TiltTwistConstraint[] TwistConstraints { get; set; } = [];
 
         /// <summary>Gets the frame cache used to retrieve and interpolate animation frames.</summary>
         public AnimationFrameCache FrameCache { get; }
@@ -88,13 +45,6 @@ namespace ValveResourceFormat.Renderer
                 field = value;
             }
         }
-
-        /// <summary>
-        /// Gets whether the active animation clip has finished playing (is not looping and has reached the end).
-        /// </summary>
-        public bool ActiveClipFinished => CurrentSubController.HasValue
-            ? CurrentSubController.Value.Handler.ActiveClipFinished
-            : activeClip != null && !activeClip.Looping && activeClip.IsPaused;
 
         /// <summary>Gets or sets the current frame index of the active animation.</summary>
         public int Frame
@@ -128,7 +78,7 @@ namespace ValveResourceFormat.Renderer
             : base(skeleton)
         {
             FrameCache = new(skeleton, flexControllers);
-            BlendedFrame = new Frame(skeleton, flexControllers);
+            BlendedFrame = new(skeleton, flexControllers);
         }
 
         /// <summary>
@@ -147,65 +97,7 @@ namespace ValveResourceFormat.Renderer
 
             if (!IsPaused && activeClip != null)
             {
-                IsPaused = clips.Values.All(c => c.IsPaused);
-                Frame = activeClip.Frame;
-
-                // Update time for all clips
-                foreach (var clip in clips.Values)
-                {
-                    if (!clip.IsPaused && clip.Animation.FrameCount > 1)
-                    {
-                        clip.Time += timeStep;
-
-                        if (!clip.Looping)
-                        {
-                            var lastFrame = clip.Animation!.FrameCount - 1;
-                            var maxTime = lastFrame / clip.Animation.Fps;
-
-                            if (clip.Time > maxTime)
-                            {
-                                clip.IsPaused = true;
-                                clip.Frame = lastFrame;
-                            }
-                        }
-                    }
-                }
-
-                if (activeClip.IsTimeBasedTransition && previousClip != null)
-                {
-                    // Distribute blend weights between previous clip and active clip only.
-                    currentBlendTime -= timeStep;
-
-                    if (currentBlendTime <= 0f)
-                    {
-                        previousClip.Weight = 0f;
-                        activeClip.Weight = 1f;
-                        previousClip = null;
-                    }
-                    else
-                    {
-                        var t = activeClip.BlendTime > 0f
-                            ? 1f - Math.Clamp(currentBlendTime / activeClip.BlendTime, 0f, 1f)
-                            : 1f;
-
-                        var blendProgress = t * t * (3f - 2f * t);
-
-                        activeClip.Weight = blendProgress;
-                        previousClip.Weight = 1f - blendProgress;
-
-                        foreach (var clip in clips.Values)
-                        {
-                            if (clip != activeClip && clip != previousClip)
-                            {
-                                clip.Weight = 0f;
-                            }
-                        }
-                    }
-
-                    var sum = clips.Values.Sum(c => c.Weight);
-                    Debug.Assert(sum > 0f, "Total blend weight should be greater than zero.");
-                    Debug.Assert(Math.Abs(sum - 1f) < 0.01f, $"Total blend weight should be approximately 1. Found: {sum}");
-                }
+                UpdateClips(timeStep);
             }
 
             if (CurrentSubController is { } subController)
@@ -216,6 +108,8 @@ namespace ValveResourceFormat.Renderer
 
                 var updated = subController.Handler.Update(timeStep);
                 IsPaused = subController.Handler.IsPaused;
+                Time = subController.Handler.Time;
+                forceUpdate = subController.Handler.forceUpdate;
 
                 if (!updated && !forceUpdate)
                 {
@@ -258,6 +152,8 @@ namespace ValveResourceFormat.Renderer
                 AnimationFrame = GetFrame();
                 updateHandler(ActiveAnimation, Frame);
                 forceUpdate = false;
+
+                ApplyInverseKinematics();
                 return true;
             }
 
@@ -295,9 +191,7 @@ namespace ValveResourceFormat.Renderer
                 GetBoneMatricesRecursive(root, Transform, AnimationFrame, Pose);
             }
 
-            ApplyConstraints();
-            ApplyFirstpersonLegs();
-
+            ApplyInverseKinematics();
             return true;
         }
 
@@ -355,95 +249,7 @@ namespace ValveResourceFormat.Renderer
 
             if (animation != null)
             {
-                var animName = animation.Name;
-
-                // Check if clip already exists
-                if (!clips.TryGetValue(animName, out var newClip))
-                {
-                    newClip = new Clip(animation) { Looping = Looping, BlendTime = blendTime };
-                    clips[animName] = newClip;
-                }
-                else
-                {
-                    // Update existing clip properties
-                    newClip.Looping = Looping;
-                    newClip.BlendTime = blendTime;
-
-                    newClip.IsPaused = false;
-                    newClip.Frame = 0;
-                }
-
-                // Handle blending
-                if (activeClip == newClip)
-                {
-                    // Re-setting the same animation should not create a self-blend transition.
-                    previousClip = null;
-
-                    foreach (var clip in clips.Values)
-                    {
-                        clip.Weight = 0f;
-                    }
-
-                    newClip.Weight = 1f;
-
-                    if (blendTime == 0f)
-                    {
-                        FrameCache.Clear();
-                    }
-                }
-                else if (blendTime > 0f && activeClip != null)
-                {
-                    // Time-based transition: only blend from previous clip -> active clip.
-                    previousClip = activeClip;
-                    previousClip.Weight = 1f;
-
-                    // Set all other clips to zero immediately.
-                    foreach (var clip in clips.Values)
-                    {
-                        if (clip != previousClip && clip != newClip)
-                        {
-                            clip.Weight = 0f;
-                        }
-                    }
-
-                    newClip.Weight = 0f;
-                    currentBlendTime = blendTime;
-                }
-                else if (blendTime == -1f && activeClip != null)
-                {
-                    // Manual blend: keep previous clip, user may set weights manually.
-                    previousClip = activeClip;
-                    previousClip.Weight = 1f;
-
-                    foreach (var clip in clips.Values)
-                    {
-                        if (clip != previousClip && clip != newClip)
-                        {
-                            clip.Weight = 0f;
-                        }
-                    }
-
-                    newClip.Weight = 0f;
-                }
-                else
-                {
-                    // No blending - disable previous clip and all other clips.
-                    previousClip = null;
-
-                    foreach (var clip in clips.Values)
-                    {
-                        clip.Weight = 0f;
-                    }
-
-                    newClip.Weight = 1f;
-
-                    if (blendTime == 0f)
-                    {
-                        FrameCache.Clear();
-                    }
-                }
-
-                activeClip = newClip;
+                TransitionToClip(animation, blendTime);
             }
             else
             {
@@ -472,72 +278,7 @@ namespace ValveResourceFormat.Renderer
                 return subController.Handler.GetFrame();
             }
 
-            if (activeClip == null)
-            {
-                return null;
-            }
-
-            // Check if blending is needed
-            var needsBlending = clips.Values.Any(c => c != activeClip && c.Weight > 0f);
-
-            if (!needsBlending)
-            {
-                return SampleFrame(activeClip);
-            }
-
-            BlendedFrame.FrameIndex = -1;
-            BlendedFrame.Bones.AsSpan().Clear();
-            BlendedFrame.Datas.AsSpan().Clear();
-
-            var totalWeight = 0f;
-            foreach (var clip in clips.Values)
-            {
-                if (clip.Weight <= 0f)
-                {
-                    continue;
-                }
-
-                var frame = SampleFrame(clip);
-                var blendFactor = clip.Weight / (totalWeight + clip.Weight);
-
-                for (var i = 0; i < frame.Bones.Length; i++)
-                {
-                    BlendedFrame.Bones[i] = BlendedFrame.Bones[i].Blend(frame.Bones[i], blendFactor);
-                }
-
-                for (var i = 0; i < frame.Datas.Length; i++)
-                {
-                    BlendedFrame.Datas[i] = float.Lerp(BlendedFrame.Datas[i], frame.Datas[i], blendFactor);
-                }
-
-                totalWeight += clip.Weight;
-            }
-
-            return BlendedFrame;
-        }
-
-        private Frame SampleFrame(Clip clip)
-        {
-            var ignoreCache = clip.Animation != ActiveAnimation;
-
-            try
-            {
-                if (ignoreCache)
-                {
-                    FrameCache.PurgeCache();
-                }
-
-                return clip.IsPaused
-                    ? FrameCache.GetFrame(clip.Animation, clip.Frame)
-                    : FrameCache.GetInterpolatedFrame(clip.Animation, clip.Time);
-            }
-            finally
-            {
-                if (ignoreCache)
-                {
-                    FrameCache.PurgeCache();
-                }
-            }
+            return GetBlendedFrame();
         }
 
         /// <summary>
@@ -619,53 +360,5 @@ namespace ValveResourceFormat.Renderer
             ExternalSkeletons[skeletonName] = new(controller, remapTable, debugMap);
         }
 
-        /// <summary>
-        /// Sets the blend weight for a clip with the specified animation name.
-        /// </summary>
-        /// <param name="name">The name of the animation.</param>
-        /// <param name="weight">The weight value (0.0 to 1.0).</param>
-        /// <param name="restartIfNew">Whether to restart the animation if its just now fading in.</param>
-        public void SetAnimationWeight(string name, float weight, bool restartIfNew = false)
-        {
-            if (clips.TryGetValue(name, out var clip))
-            {
-                var wasZero = clip.Weight == 0f;
-                clip.Weight = weight;
-
-                if (restartIfNew && wasZero && weight > 0f)
-                {
-                    clip.Time = 0f;
-                    clip.IsPaused = false;
-                }
-            }
-
-            if (CurrentSubController is { } subController)
-            {
-                subController.Handler.SetAnimationWeight(name, weight, restartIfNew);
-            }
-        }
-
-
-        public void SetAnimationProperties(string name, float? time = null, bool? looping = null)
-        {
-            if (clips.TryGetValue(name, out var clip))
-            {
-                if (time.HasValue)
-                {
-                    clip.Time = time.Value;
-                    clip.IsPaused = false;
-                }
-
-                if (looping.HasValue)
-                {
-                    clip.Looping = looping.Value;
-                }
-            }
-
-            if (CurrentSubController is { } subController)
-            {
-                subController.Handler.SetAnimationProperties(name, time, looping);
-            }
-        }
     }
 }
